services/data_storage_service/data_storage_service/requirements/base.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy[asyncio]
asyncpg
pydantic
redis
httpx
loguru
python-decouple
alembic
pydantic-settings
structlog